Fix issue with pagination on mobile
On mobile, the right panel of the footer would be on top of the
paginator, blocking it from receiving click events.  Fixed by setting
the width back to auto below the break size.  Width needs to be fixed
above the break size to allow the left and right panels to be set the
same width in order to properly center the center panel.